Pantalla de acceso no autorizado y usuario logeado no se puede eliminar

+++ src/main/java/com/gl05/bad/servicio/UserServiceImp.java
@@ -6,6 +6,8 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.jpa.datatables.mapping.DataTablesInput;
 import org.springframework.data.jpa.datatables.mapping.DataTablesOutput;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -16,7 +18,7 @@ public class UserServiceImp implements UserService {
     private UsuarioDao usuarioDao;
 
     @Override
-    @Transactional(readOnly=true)
+    @Transactional(readOnly = true)
     public List<Usuario> listaUsuarios() {
         return (List<Usuario>) usuarioDao.findAll();
     }
@@ -30,24 +32,35 @@ public class UserServiceImp implements UserService {
     @Override
     @Transactional
     public void eliminarUsuario(Usuario usuario) {
+        // Obtener el usuario iniciado en sesión
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String username = authentication.getName();
+        Usuario usuarioActual = usuarioDao.findByUsername(username);
+
+        // Verificar si el usuario a eliminar es el usuario iniciado en sesión
+        if (usuarioActual.getIdUsuario().equals(usuario.getIdUsuario())) {
+            throw new IllegalArgumentException("No se puede eliminar el usuario iniciado en sesión.");
+        }
+
+        // Proceder con la eliminación del usuario
         usuarioDao.delete(usuario);
     }
 
     @Override
-    @Transactional(readOnly=true)
+    @Transactional(readOnly = true)
     public Usuario encontrarUsuario(Long idUsuario) {
         return usuarioDao.findById(idUsuario).orElse(null);
     }
-    
+
     @Override
-    @Transactional(readOnly=true)
-    public Usuario encontrarUsuarioPorUsername(String username){
+    @Transactional(readOnly = true)
+    public Usuario encontrarUsuarioPorUsername(String username) {
         return usuarioDao.findByUsername(username);
     }
-    
+
     @Override
-    @Transactional(readOnly=true)
-    public Usuario encontrarUsuarioPorEmail(String email){
+    @Transactional(readOnly = true)
+    public Usuario encontrarUsuarioPorEmail(String email) {
         return usuarioDao.findByEmail(email);
     }
 
@@ -63,9 +76,9 @@ public class UserServiceImp implements UserService {
     }
 
     @Override
-    @Transactional(readOnly=true)
+    @Transactional(readOnly = true)
     public DataTablesOutput<Usuario> listarUsuarios(DataTablesInput input) {
-        return (DataTablesOutput<Usuario>)usuarioDao.findAll(input);
+        return (DataTablesOutput<Usuario>) usuarioDao.findAll(input);
     }
 
 }
